tests: write tests for macro `concat_const_str`, move `const_data` generic_test to tests for `const_data`
